refactor(memory): extract view utilities
