+++ src/components/familiars-memory-view-compact-path.test.ts
@@ -2,7 +2,7 @@
 import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 
-const source = await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8");
+const source = await readFile(new URL("./familiars-memory-utils.ts", import.meta.url), "utf8");
 
 // Static-source assertions: the function must define a threshold + take the last 3 segments.
 assert.match(
