+++ src/components/familiars-memory-view-detail.test.ts
@@ -2,7 +2,10 @@
 import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
 
-const source = await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8");
+const source = [
+  await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8"),
+  await readFile(new URL("./familiars-memory-utils.ts", import.meta.url), "utf8"),
+].join("\n");
 const hook = await readFile(new URL("../lib/use-memory-file.ts", import.meta.url), "utf8");
 assert.match(hook, /path=\$\{encodeURIComponent\(path\)\}/,
   "the shared hook must URL-encode the requested path");
