+++ src/components/familiars-memory-view.tsx
@@ -33,24 +33,9 @@ import { SkeletonRows } from "@/components/ui/skeleton";
 import { Button } from "@/components/ui/button";
 import { StandardSelect } from "@/components/ui/select";
 import { MemoryReaderPane } from "@/components/familiars-memory-reader";
+import { compactPath, fileBase, fileDir, formatBytes, memoryMatches, type FileMemoryEntry } from "@/components/familiars-memory-utils";
 
-export type FileMemoryEntry = {
-  root: string;
-  rootLabel: string;
-  relPath: string;
-  fullPath: string;
-  size: number;
-  modified: string;
-  sourceId: string;
-  sourceKind: "coven-origin" | "external-harness" | "runtime";
-  sourceKindLabel: string;
-  rootPath: string;
-  origin?: "coven";
-  harnessId?: string;
-  runtimeId?: string;
-  sourceContext?: string;
-  familiarId?: string;
-};
+export type { FileMemoryEntry } from "@/components/familiars-memory-utils";
 
 /**
  * Memory data supplied by a parent that already fetches /api/coven-memory +
@@ -89,60 +74,6 @@ type CovenMemoryResponse =
 type FileMemoryResponse =
   | { ok: true; entries: FileMemoryEntry[] }
   | { ok: false; entries?: FileMemoryEntry[]; error?: string };
-
-function compactPath(path: string): string {
-  const collapsed = path.replace(/^\/Users\/[^/]+/, "~");
-  const THRESHOLD = 52;
-  if (collapsed.length <= THRESHOLD) return collapsed;
-  const segments = collapsed.split("/").filter(Boolean);
-  if (segments.length <= 4) return collapsed;
-  const first = collapsed.startsWith("~") ? "~" : `/${segments[0]}`;
-  const last = segments.slice(-3);
-  return `${first}/…/${last.join("/")}`;
-}
-
-function fileBase(p: string): string {
-  const segments = p.split("/").filter(Boolean);
-  return segments[segments.length - 1] ?? p;
-}
-
-/** Directory portion of a full path, collapsed to ~ and ellipsized. "" when at root. */
-function fileDir(fullPath: string): string {
-  const base = fileBase(fullPath);
-  const parent = fullPath.slice(0, Math.max(0, fullPath.length - base.length)).replace(/\/$/, "");
-  return parent ? compactPath(parent) : "";
-}
-
-function formatBytes(n: number | undefined): string {
-  if (!n || n < 0 || !Number.isFinite(n)) return "";
-  if (n < 1024) return `${n} B`;
-  if (n < 1024 * 1024) return `${(n / 1024).toFixed(n < 10 * 1024 ? 1 : 0)} KB`;
-  return `${(n / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-function memoryMatches(entry: CovenMemoryEntry | FileMemoryEntry, query: string): boolean {
-  if (!query) return true;
-  if ("familiar_id" in entry) {
-    return [
-      entry.title,
-      entry.excerpt ?? "",
-      entry.familiar_id,
-      entry.path,
-      entry.source_context ?? "",
-    ].some((value) => value.toLowerCase().includes(query));
-  }
-  return [
-    entry.rootLabel,
-    entry.sourceKindLabel,
-    entry.harnessId ?? "",
-    entry.runtimeId ?? "",
-    entry.origin ?? "",
-    entry.familiarId ?? "",
-    entry.relPath,
-    entry.fullPath,
-    entry.sourceContext ?? "",
-  ].some((value) => value.toLowerCase().includes(query));
-}
 
 export function FamiliarsMemoryView({ familiars, activeFamiliar, onOpenMemoryFile, limit, lockToFamiliar, compact, feed }: Props) {
   useDateTimePrefs(); // subscribe: re-render when the date/time density pref changes
